Creating endpoint to update a service
I created this endpoint to update a service.
The only attribute that can be updated is the service value.
That way, will be passed an ID, the API will search for a service that has that ID. If the API found, its value is updated. If not, an exception is thrown