Creating the unit test of the save method of the "Servico" service
I created the two scenarios that can occur when this endpoint is called.
The first scenary is the success. The service is saved in database normally.
The second scenary is when the service being saved already exists in database (checking is done by name). In this case, an exception is thrown.