Creating utility classes
I created these two classes and two both interfaces.
They will be used by the API in a polymorphic way. For example: The "ValidacaoDados" interface contains a "validar" method. I will create several subclasses that extend this interface so that each one implements the "validar" method based on its logic. The "VeiculoValidacao" subclass implements the method to validate the vehicle's license plate, the "PessoaValidacao" class will implement the method validating a person's data.
The same applies to the "PersistenciaDados" interface.