Creating the method to search for a record by an attribute
This method allows the user to pass an attribute that extends Object and search for a record in the database using this attribute.
In the example, the car plate is passed as an Object and converted to a String. The search is done and a Veiculo is returned or an exception is thrown.