Creating the integration test of the findByPlaca endpoint of the VeiculoRestController
This two tests validates two requests made to the endpoint. The first tests a request by receiving a plate that exists in the database, then the API returns a succesful response.
The second tests a request with a plate that does not exist in the database. An exception will be thrown in this case.