Creating the update endpoint of the VeiculoRestController
I created the endpoint that the user calls to update a vehicle's data.
When called, the API will search for a vehicle in the database that has the same ID passed. If found, the vehicle's data will be updated and it will be saved again in the database. If not, the API throws a 404 NOT FOUND.