Creating the class to update object's data
The "atualizarDados" method will be called whenever an object needs to have its data replaced with updated data.
With the update endpoint, an object with updated data is passed in the request and an object in the database receives this data. This method will be called to make this replacement.