Testing the update endpoint of the VeiculoRestController
I validated both test cases for this endpoint. The first is when the ID passed in the request is valid. The API will search for a record in the database that contains this ID and will replace its data.
The second is when the ID is invalid. In this case, the API will throw a 404 NOT FOUND.